refactor(Notice, Wish)! : enum 리펙터링
### 변경점
- 프론트에서 int로 입력해도 문제 없도록 변환단계 추가

+++ src/main/java/sesac/bookmanager/notice/data/NoticeType.java
@@ -23,9 +23,9 @@ public enum NoticeType {
     }
 
     @JsonCreator
-    public static NoticeType fromCode(byte code) {
+    public static NoticeType fromCode(int code) {
         for (NoticeType noticeType : NoticeType.values()) {
-            if (noticeType.code == code) {
+            if (noticeType.code == (byte) code) {
                 return noticeType;
             }
         }
